Organize Mongo Helper Functions
