+++ src/shared/Mongo/HelperFunctions/index.ts
@@ -1,0 +1,15 @@
+import { documentLearningObject } from './documentLearningObject/documentLearningObject';
+import { documentReleasedLearningObject } from './documentReleasedLearningObject/documentReleasedLearningObject';
+import { generateLearningObject } from './generateLearningObject/generateLearningObject';
+import { generateLearningObjectSummary } from './generateLearningObjectSummary/generateLearningObjectSummary';
+import { generateReleasedLearningObject } from './generateReleasedLearningObject/generateReleasedLearningObject';
+import { generateReleasedLearningObjectSummary } from './generateReleasedLearningObjectSummary/generateReleasedLearningObjectSummary';
+
+export {
+    documentLearningObject,
+    documentReleasedLearningObject,
+    generateLearningObject,
+    generateLearningObjectSummary,
+    generateReleasedLearningObject,
+    generateReleasedLearningObjectSummary,
+ };
