+++ src/shared/functions/isEmail/isEmail.ts
@@ -1,0 +1,7 @@
+export function isEmail(value: string): boolean {
+    const emailPattern = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+    if (emailPattern.test(value)) {
+      return true;
+    }
+    return false;
+  }
